Sprite class build

+++ CoinCatcher/Game Logic/Basic Logic/Game.cpp
@@ -3,10 +3,10 @@
 
 Game::Game() {
 	ResourceManager::LoadShader("Sprite.vs", "sprite.fs", nullptr, "standard");
-	ResourceManager::LoadTexture("Resources/Sprite/player.bmp", false, "player");
+	ResourceManager::LoadTexture("Resources/Sprite/dino.png", true, "player");
 	ResourceManager::LoadTexture("Resources/Sprite/coin.png", true, "coin");
 	
-	player = new Player(100, 100, "player" , 10, 5);
+	player = new Player(100, 100, "player", "standard", 10, 5);
 	window = &Window::getInstance();
 } 
 
@@ -23,7 +23,7 @@ void Game::setGameMode(unsigned int mode)
 
 // found in OpenGL Game Development by Example just rewrote to be compatible with this program 
 void Game::SpawnCoin() {
-	Coin* coin = new Coin(70, 70, "coin", 10, 10, 1);
+	Coin* coin = new Coin(70, 70, "coin", "standard", 10, 10, 1);
 
 	float marginX = coin->getWidth();
 	float marginY = coin->getHeight();
@@ -51,9 +51,8 @@ void Game::timer(void(*t)(int)) {
 
 void Game::Draw() {
 	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-	cout << "Draw Start\n";
+
 	if (player != nullptr) {
-		cout << "player";
 		player->Draw();
 	}
 	for(auto coin : coins)
@@ -100,7 +99,6 @@ void Game::DeleteCoins(vector<int>& needDeletedIndex) {
 void Game::PauseGame() {
 }
 void Game::GameLoop() {
-	Draw();
 
 	if (rand() % 100 > 30)
 		SpawnCoin();
@@ -118,7 +116,10 @@ void Game::GameLoop() {
 			//deletedIndex.push_back(i);
 		}
 	}
-	DeleteCoins(deletedIndex);
+	if(!deletedIndex.empty())
+		DeleteCoins(deletedIndex);
+
+	Draw();
 
 	// UI
 }
